版本：v0.1.0   1、完善ui   2、开始写同步线程，采用信号与槽进行线程间的数据沟通，定义了一部分信号。

+++ src/test-baiduSyncGUI.py
@@ -7,6 +7,7 @@ import bypy
 import configparser as config
 import os
 import uiobj.configWin as configWin
+from threading import Thread,Lock
 
 
 configini_dir='src/config.ini'
@@ -19,15 +20,25 @@ class SystemTray(object):
         self.app = app
         self.cfgw = cfgw
         QApplication.setQuitOnLastWindowClosed(False)  # 禁止默认的closed方法，只能使用qapp.quit()的方法退出程序
-
         self.tp = QSystemTrayIcon(self.cfgw)
         self.initUI()
         self.run()
 
     def initUI(self):
         # 设置托盘图标
-
         self.tp.setIcon(QIcon('./res/cloud.ico'))
+        #初始化bypy
+        self.mybp=bypy.ByPy()
+        infomsg=self.mybp.info()
+        print(infomsg)
+        #检查本地与远程目录正确性
+        self.syncFlag=1      #启用同步，如果是0则是不启用同步
+        if not os.path.exists(self.cfgw.localPath):
+            QMessageBox.question(self.cfgw,"提示","不存在本地文件夹，请设置本地同步目录")
+            self.syncFlag=0
+        if self.mybp.meta(self.cfgw.remotePath)!=0 :
+            QMessageBox.question(self.cfgw,"提示","不存在远程文件夹，请设置远程同步目录")
+            self.syncFlag=0
 
     def quitApp(self):
         # 退出程序
@@ -46,21 +57,25 @@ class SystemTray(object):
         # 主界面显示方法
         # 鼠标点击icon传递的信号会带有一个整形的值，1是表示单击右键，2是双击，3是单击左键，4是用鼠标中键点击
         if reason == 2:
-            rn=os.system("xdg-open "+self.cfgw.localPath)
-            print(rn)
-            #self.cfgw.show()
+            self.xdgopenFile()
+
+    def xdgopenFile(self):
+        rn=os.system("xdg-open "+self.cfgw.localPath)
+        print(rn)
+       
 
     def run(self):
 
         a1 = QAction('&设置(Show)', triggered=self.cfgw.show)
-        a2 = QAction('&退出(Exit)', triggered=self.quitApp)
-
+        a2 = QAction('&显示本地文件',triggered=self.xdgopenFile)
+        aq = QAction('&退出(Exit)', triggered=self.quitApp)
+        
         tpMenu = QMenu()
         tpMenu.addAction(a1)
         tpMenu.addAction(a2)
+        tpMenu.addAction(aq)
         self.tp.setContextMenu(tpMenu)
         self.tp.show()  # 不调用show不会显示系统托盘消息，图标隐藏无法调用
-
         # 信息提示
         # 参数1：标题
         # 参数2：内容
@@ -70,9 +85,49 @@ class SystemTray(object):
         self.tp.messageClicked.connect(self.message)
         # 绑定托盘菜单点击事件
         self.tp.activated.connect(self.act)
+
+        #开启同步线程
+
+
         sys.exit(self.app.exec_())  # 持续对app的连接
 
+   
+class syncThread(QtCore.QThread):
+    #同步线程，负责进行实时的同步
+    #由同步线程发出的信号，告诉主线程同步状态
+    syncStatus_sign=QtCore.pyqtSignal(str)
+    def __init__(self,localPath,remotePath,syncTime,syncFlag):
+        super(syncThread,self).__init__()
+        #关键参数，由主线程告诉同步线程
+        self.localPath=localPath                                  #本地目录
+        self.remotePath=remotePath                        #远程目录
+        self.syncTime=syncTime                                 #自动同步时间
+        self.syncFlag=syncFlag 
+        self.mybp=bypy.ByPy()                                   #是否同步的标志位
+        self.exitflg=1                                                        #退出线程标志位，0时退出线程
 
+    #槽函数，由主线程发出信号触发的函数
+    def changFLAG(self,flg):
+        self.syncFlag=flg
+    def changCONFIG(self,cfglist):
+        self.localPath=cfglist[0]
+        self.remotePath=cfglist[1]
+        self.syncTime=cfglist[2]
+    def syncupNOW(self):
+        #只要有信号过来，立马同步上传
+        self.mybp.syncup(self.localPath,self.remotePath,True)
+
+    #线程函数
+    def run(self):
+        #首先程序开始，先把云端内容下下来
+        if self.syncFlag==1:
+            self.mybp.syncdown(self.remotePath,self.localPath,True)
+        #定时上传
+        while self.exitflg:
+            if self.syncFlag==1:
+                self.mybp.syncup(self.localPath,self.remotePath,True)
+            
+            self.sleep(60*self.syncTime)
 
 
 class configWindows(QWidget,configWin.Ui_Form):
@@ -80,7 +135,6 @@ class configWindows(QWidget,configWin.Ui_Form):
     def __init__(self):
         super(configWindows,self).__init__()
         self.setupUi(self)
-        
         #读取配置文件
         self.cfg = config.ConfigParser()
         self.cfg.read(configini_dir)
@@ -89,16 +143,25 @@ class configWindows(QWidget,configWin.Ui_Form):
         self.syncTime=int(self.cfg.get("SET","syncTime"))
         print("读取配置文件\n本地："+self.localPath+"\n远程："+self.remotePath)
         print("同步时间：",self.syncTime)
-        #更新设置界面的显示
-        self.updatePathshow()
         #连接信号与槽
         self.pushButton.clicked.connect(self.On_selectPath)
         self.pushButton_2.clicked.connect(self.On_saveConfig)
+        #自定义信号，由主线程发送/由设置窗口发送
+        self.config_sign=QtCore.pyqtSignal(list)                     #配置参数信号，发送该信号可以改变同步线程里的配置参数
+        #初始化comboBox
+        self.syncTime_cBoxlist=['1','2','3','4','5','7','10','20','30']
+        self.comboBox.addItems(self.syncTime_cBoxlist)
+         #更新设置界面的显示
+        self.updateShow()
         
-    def updatePathshow(self):
+        
+        
+        
+    def updateShow(self):
         # 更新设置界面的显示
         self.lineEdit_2.setText(self.localPath)
         self.lineEdit.setText(self.remotePath)
+        self.comboBox.setCurrentIndex(self.syncTime_cBoxlist.index(str(self.syncTime)))
 
     def On_selectPath(self):
         #槽、回调函数:选择本地文件夹
@@ -106,21 +169,31 @@ class configWindows(QWidget,configWin.Ui_Form):
         if len(local_dir)!=0 :
             self.lineEdit_2.setText(local_dir)
     def On_saveConfig(self):
+         #检查本地与远程目录正确性
+        if not os.path.exists(self.lineEdit_2.text()):
+            QMessageBox.question(self,"提示","不存在本地文件夹，请重新设置本地同步目录")
+            return
         #槽、回调函数:保存设置
         self.localPath=self.lineEdit_2.text()
         self.remotePath=self.lineEdit.text()
+        self.syncTime=int(self.comboBox.currentText())
         self.cfg.set('SET','localPath',self.localPath)
         self.cfg.set('SET','remotePath',self.remotePath)
+        self.cfg.set('SET','syncTime',str(self.syncTime))
         self.cfg.write(open(configini_dir, "w"))
         print("应用成功")
- 
+        print("localPath",type(self.localPath),self.localPath)
+        print("remotePath",type(self.remotePath),self.remotePath)
+        print("syncTime",type(self.syncTime),self.syncTime)
+    
+    def closeEvent(self, event):
+        self.updateShow()
 
 if __name__ == "__main__":
     QtCore.QCoreApplication.setAttribute(QtCore.Qt.AA_EnableHighDpiScaling)
 
     # 创建一个app程序
     app = QApplication(sys.argv)
-    
     # 创建窗口
     #win = Window()
     configww=configWindows()
